Move FileList interfaces to types folder
